Revert "OLH-2720 Resolve access denied error that is being logged"
