+++ sonar-project.properties
@@ -5,7 +5,7 @@ sonar.projectName=di-account-management-frontend
 sonar.host.url=https://sonarcloud.io
 
 sonar.sources=src/
-sonar.exclusions=src/**/*.test.ts,src/assets/javascript/*.js,src/app.ts, src/server.ts, src/config.ts, post-deploy-tests/**/*, src/components/oidc-callback/call-back-controller.ts
+sonar.exclusions=src/**/*.test.ts,src/assets/javascript/*.js,src/app.ts, src/server.ts, src/config.ts, post-deploy-tests/**/*
 sonar.tests=test/,src/
 sonar.test.inclusions=**/*.test.ts
