+++ src/components/backchannel-logout/backchannel-logout-routes.ts
@@ -1,13 +1,14 @@
 import * as express from "express";
 import { backchannelLogoutPost } from "./backchannel-logout-controller";
 import { PATH_DATA } from "../../app.constants";
+import { asyncHandler } from "../../utils/async";
 import { globalTryCatchAsync } from "../../utils/global-try-catch";
 
 const router = express.Router();
 
 router.post(
   PATH_DATA.BACKCHANNEL_LOGOUT.url,
-  globalTryCatchAsync(backchannelLogoutPost)
+  globalTryCatchAsync(asyncHandler(backchannelLogoutPost))
 );
 
 export { router as backchannelLogoutRouter };
