+++ src/components/change-email/change-email-routes.ts
@@ -1,6 +1,7 @@
 import { PATH_DATA } from "../../app.constants";
 import { changeEmailPost, changeEmailGet } from "./change-email-controller";
 import * as express from "express";
+import { asyncHandler } from "../../utils/async";
 import { validateChangeEmailRequest } from "./change-email-validation";
 import { requiresAuthMiddleware } from "../../middleware/requires-auth-middleware";
 import { validateStateMiddleware } from "../../middleware/validate-state-middleware";
@@ -22,7 +23,7 @@ router.post(
   validateStateMiddleware,
   validateChangeEmailRequest(),
   refreshTokenMiddleware(),
-  globalTryCatchAsync(changeEmailPost())
+  globalTryCatchAsync(asyncHandler(changeEmailPost()))
 );
 
 export { router as changeEmailRouter };
