+++ src/components/change-password/change-password-routes.ts
@@ -5,6 +5,7 @@ import {
   changePasswordPost,
 } from "./change-password-controller";
 import { validateChangePasswordRequest } from "./change-password-validation";
+import { asyncHandler } from "../../utils/async";
 import { requiresAuthMiddleware } from "../../middleware/requires-auth-middleware";
 import { validateStateMiddleware } from "../../middleware/validate-state-middleware";
 import { refreshTokenMiddleware } from "../../middleware/refresh-token-middleware";
@@ -28,7 +29,7 @@ router.post(
   validateStateMiddleware,
   validateChangePasswordRequest(),
   refreshTokenMiddleware(),
-  globalTryCatchAsync(changePasswordPost())
+  globalTryCatchAsync(asyncHandler(changePasswordPost()))
 );
 
 export { router as changePasswordRouter };
