+++ src/components/enter-password/enter-password-routes.ts
@@ -10,6 +10,7 @@ import {
   globalTryCatch,
   globalTryCatchAsync,
 } from "../../utils/global-try-catch";
+import { asyncHandler } from "../../utils/async";
 
 const router = express.Router();
 
@@ -23,7 +24,7 @@ router.post(
   PATH_DATA.ENTER_PASSWORD.url,
   requiresAuthMiddleware,
   refreshTokenMiddleware(),
-  globalTryCatchAsync(enterPasswordPost())
+  globalTryCatchAsync(asyncHandler(enterPasswordPost()))
 );
 
 export { router as enterPasswordRouter };
