+++ src/components/oidc-callback/tests/callback-controller.test.ts
@@ -27,7 +27,6 @@ describe("callback controller", () => {
         user: {},
         destroy: sandbox.fake(),
         state: sandbox.fake(),
-        nonce: sandbox.fake(),
       } as any,
       t: sandbox.fake(),
       oidc: {
@@ -223,27 +222,6 @@ describe("callback controller", () => {
 
       await oidcAuthCallbackGet(fakeService)(req as Request, res as Response);
 
-      expect(res.redirect).to.have.calledWith(PATH_DATA.SESSION_EXPIRED.url);
-    });
-
-    it("should redirect to session expired if session nonce is missing", async () => {
-      req.session.nonce = undefined;
-
-      const fakeService: ClientAssertionServiceInterface = {
-        generateAssertionJwt: sandbox.fake(),
-      };
-
-      await oidcAuthCallbackGet(fakeService)(req as Request, res as Response);
-
-      expect(res.redirect).to.have.calledWith(PATH_DATA.SESSION_EXPIRED.url);
-    });
-
-    it("redirect to session expired when access denied error is thrown", async () => {
-      req.oidc.callbackParams = sandbox.fake.throws(new Error("access_denied"));
-      const fakeService: ClientAssertionServiceInterface = {
-        generateAssertionJwt: sandbox.fake.resolves("testassert"),
-      };
-      await oidcAuthCallbackGet(fakeService)(req as Request, res as Response);
       expect(res.redirect).to.have.calledWith(PATH_DATA.SESSION_EXPIRED.url);
     });
   });
