--- src/components/oidc-callback/tests/callback-utils.test.ts
@@ -1,243 +1,0 @@
-import { expect } from "chai";
-import sinon from "sinon";
-import { describe } from "mocha";
-import {
-  exchangeToken,
-  determineRedirectUri,
-  COOKIE_CONSENT,
-  handleOidcCallbackError,
-  populateSessionWithUserInfo,
-  attachSessionIdsFromGsCookie,
-} from "../call-back-utils";
-import { PATH_DATA } from "../../../app.constants";
-import * as sessionStore from "../../../utils/session-store";
-import { logger } from "../../../utils/logger";
-
-describe("callback-utils", () => {
-  afterEach(() => {
-    sinon.restore();
-  });
-
-  describe("exchangeToken", () => {
-    it("should call req.oidc.callback with expected arguments", async () => {
-      const fakeTokenSet = { access_token: "abc", id_token: "xyz" };
-
-      const req: any = {
-        session: { state: "test-state", nonce: "test-nonce" },
-        oidc: {
-          metadata: {
-            client_id: "client123",
-            redirect_uris: ["http://localhost/callback"],
-          },
-          issuer: {
-            metadata: {
-              token_endpoint: "http://issuer/token",
-            },
-          },
-          callback: sinon.fake.resolves(fakeTokenSet),
-        },
-      };
-
-      const service = {
-        generateAssertionJwt: sinon.fake.resolves("jwt-token"),
-      };
-
-      const result = await exchangeToken(req, service as any, {
-        code: "auth-code",
-      });
-
-      expect(result).to.equal(fakeTokenSet);
-      expect(
-        service.generateAssertionJwt.calledWith(
-          "client123",
-          "http://issuer/token"
-        )
-      ).to.be.true;
-      expect(req.oidc.callback.calledOnce).to.be.true;
-    });
-  });
-
-  describe("determineRedirectUri", () => {
-    it("should return currentURL from session if available", () => {
-      const req: any = {
-        session: { currentURL: "/dashboard" },
-        query: {},
-      };
-      const res: any = {};
-      const result = determineRedirectUri(req, res);
-      expect(result).to.equal("/dashboard");
-    });
-
-    it("should return default path if currentURL not set", () => {
-      const req: any = {
-        session: {},
-        query: {},
-      };
-      const res: any = {};
-      const result = determineRedirectUri(req, res);
-      expect(result).to.equal(PATH_DATA.YOUR_SERVICES.url);
-    });
-
-    it("should append _ga param to redirect URL if consent is accepted", () => {
-      const req: any = {
-        session: {},
-        query: {
-          cookie_consent: COOKIE_CONSENT.ACCEPT,
-          _ga: "GA1.2.123456",
-        },
-      };
-      const res: any = {
-        cookie: sinon.fake(),
-        locals: {
-          trace: "fake_trace",
-          analyticsCookieDomain: "fake_analytics_domain",
-        },
-      };
-      const result = determineRedirectUri(req, res);
-      expect(result).to.equal(
-        `${PATH_DATA.YOUR_SERVICES.url}?_ga=GA1.2.123456`
-      );
-    });
-
-    it("should not append _ga if consent is not ACCEPT", () => {
-      const req: any = {
-        session: {},
-        query: {
-          cookie_consent: COOKIE_CONSENT.REJECT,
-          _ga: "GA1.2.123456",
-        },
-      };
-      const res: any = {
-        cookie: sinon.fake(),
-        locals: {
-          trace: "fake_trace",
-          analyticsCookieDomain: "fake_analytics_domain",
-        },
-      };
-      const result = determineRedirectUri(req, res);
-      expect(result).to.equal(PATH_DATA.YOUR_SERVICES.url);
-    });
-  });
-
-  describe("handleOidcCallbackError", () => {
-    it("clears session and redirects to SESSION_EXPIRED", async () => {
-      const deleteExpressSessionStub = sinon.stub(
-        sessionStore,
-        "deleteExpressSession"
-      );
-      const loggerWarnStub = sinon.stub(logger, "warn");
-
-      const req: any = { session: {}, oidc: {}, cookies: {} };
-      const res: any = {
-        locals: { trace: "trace-id" },
-        redirect: sinon.fake(),
-      };
-      const queryParams = {
-        error: "access_denied",
-        error_description: "User denied access",
-      };
-
-      await handleOidcCallbackError(req, res, queryParams);
-
-      expect(loggerWarnStub.calledOnce).to.be.true;
-      expect(deleteExpressSessionStub.calledWith(req)).to.be.true;
-      expect(res.redirect.calledWith(PATH_DATA.SESSION_EXPIRED.url)).to.be.true;
-    });
-  });
-
-  describe("populateSessionWithUserInfo", () => {
-    it("stores user info and tokens in session", () => {
-      const req: any = {
-        session: {},
-      };
-      const res: any = {
-        locals: {},
-      };
-
-      const userInfo = {
-        email: "user@example.com",
-        phone_number: "1234567890",
-        phone_number_verified: true,
-        sub: "subject123",
-        legacy_subject_id: "legacy123",
-        public_subject_id: "public123",
-      };
-
-      const tokenSet = {
-        id_token: "id.token",
-        access_token: "access.token",
-        refresh_token: "refresh.token",
-      };
-
-      populateSessionWithUserInfo(req, res, userInfo as any, tokenSet as any);
-
-      expect(req.session.user.email).to.equal("user@example.com");
-      expect(req.session.user.tokens.accessToken).to.equal("access.token");
-      expect(req.session.user.legacySubjectId).to.equal("legacy123");
-      expect(res.locals.isUserLoggedIn).to.be.true;
-      expect(req.session.user_id).to.equal("subject123");
-    });
-  });
-
-  describe("attachSessionIdsFromGsCookie", () => {
-    it("logs warning if gs cookie is malformed", () => {
-      const loggerInfo = sinon.stub(logger, "info");
-      const loggerError = sinon.stub(logger, "error");
-
-      const req: any = {
-        cookies: {
-          gs: "malformed-cookie",
-        },
-        session: {},
-      };
-
-      const res: any = {
-        locals: { trace: "abc123" },
-      };
-
-      attachSessionIdsFromGsCookie(req, res);
-
-      expect(loggerInfo.called).to.be.true;
-      expect(loggerError.calledOnce).to.be.true;
-      expect(req.session.authSessionIds).to.be.undefined;
-    });
-
-    it("sets session.authSessionIds when cookie is valid", () => {
-      const loggerInfo = sinon.stub(logger, "info");
-      const req: any = {
-        cookies: {
-          gs: "sess123.client456",
-        },
-        session: {},
-      };
-
-      const res: any = {
-        locals: { trace: "trace-xyz" },
-      };
-
-      attachSessionIdsFromGsCookie(req, res);
-
-      expect(loggerInfo.calledOnce).to.be.true;
-      expect(req.session.authSessionIds).to.deep.equal({
-        sessionId: "sess123",
-        clientSessionId: "client456",
-      });
-    });
-
-    it("logs info if gs cookie is missing", () => {
-      const loggerInfo = sinon.stub(logger, "info");
-
-      const req: any = {
-        cookies: {},
-        session: {},
-      };
-      const res: any = {
-        locals: { trace: "trace-abc" },
-      };
-
-      attachSessionIdsFromGsCookie(req, res);
-
-      expect(loggerInfo.calledOnce).to.be.true;
-    });
-  });
-});
