+++ src/utils/global-try-catch.ts
@@ -1,18 +1,24 @@
-import { NextFunction, Response, Request, RequestHandler } from "express";
+import { NextFunction, Response, Request } from "express";
 import { logger } from "./logger";
 import { MetricUnit } from "@aws-lambda-powertools/metrics";
 
-export function globalTryCatchAsync(
-  fn: RequestHandler
-): (req: Request, res: Response, next: NextFunction) => Promise<void> {
-  return function (req: Request, res: Response, next: NextFunction) {
-    return Promise.resolve(fn(req, res, next)).catch((error) => {
+export const globalTryCatchAsync = (
+  fn: (req: Request, res: Response, next?: NextFunction) => Promise<void>
+) => {
+  return async (
+    req: Request,
+    res: Response,
+    next?: NextFunction
+  ): Promise<void> => {
+    try {
+      await fn(req, res, next);
+    } catch (error) {
       req.metrics?.addMetric("globalTryCatchAsyncError", MetricUnit.Count, 1);
-      logger.error(`Global async error handler : ${error}`);
+      logger.error(`Global try catch Async: failed with the error ${error}`);
       next?.(error);
-    });
+    }
   };
-}
+};
 
 export const globalTryCatch = (
   fn: (req: Request, res: Response, next?: NextFunction) => void
@@ -22,7 +28,7 @@ export const globalTryCatch = (
       fn(req, res, next);
     } catch (error) {
       req.metrics?.addMetric("globalTryCatchError", MetricUnit.Count, 1);
-      logger.error(`Global error handler : ${error}`);
+      logger.error(`Global try catch: failed with the error ${error}`);
       next?.(error);
     }
   };
